Update npm dependencies and Fix next/head errors
Update npm dependencies to use latest and greatest

Fix the next/head errors that occur by using google tags in _document.js

Make Changes to code for accounting to new npm dependencies

Layout framework for implementing custom expiry date
