+++ frontend/package.json
@@ -1,43 +1,43 @@
 {
-  "name": "frontend",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "axios": "^0.21.1",
-    "next": "^10.1.3",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-dropdown": "^1.7.0",
-    "react-katex": "^2.0.2",
-    "react-markdown": "^4.3.1",
-    "react-modal": "^3.11.2",
-    "react-simple-code-editor": "^0.11.0",
-    "react-syntax-highlighter": "^12.2.1",
-    "react-syntax-highlighter-virtualized-renderer": "^1.1.0",
-    "styled-components": "^5.1.0",
-    "use-clipboard-copy": "^0.1.2"
-  },
-  "scripts": {
-    "dev": "next dev",
-    "start": "next start",
-    "build": "next build"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "devDependencies": {
-    "babel-plugin-styled-components": "^1.12.0"
-  }
+    "name": "frontend",
+    "version": "0.1.0",
+    "private": true,
+    "dependencies": {
+        "axios": "^1.4.0",
+        "next": "^13.4.7",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "react-dropdown": "^1.11.0",
+        "react-katex": "^3.0.1",
+        "react-markdown": "^8.0.7",
+        "react-modal": "^3.16.1",
+        "react-simple-code-editor": "^0.13.1",
+        "react-syntax-highlighter": "^15.5.0",
+        "react-syntax-highlighter-virtualized-renderer": "^1.1.0",
+        "styled-components": "^6.0.1",
+        "use-clipboard-copy": "^0.2.0"
+    },
+    "scripts": {
+        "dev": "next dev",
+        "start": "next start",
+        "build": "next build"
+    },
+    "eslintConfig": {
+        "extends": "react-app"
+    },
+    "browserslist": {
+        "production": [
+            ">0.2%",
+            "not dead",
+            "not op_mini all"
+        ],
+        "development": [
+            "last 1 chrome version",
+            "last 1 firefox version",
+            "last 1 safari version"
+        ]
+    },
+    "devDependencies": {
+        "babel-plugin-styled-components": "^2.1.4"
+    }
 }
