+++ frontend/src/components/Inputs/Dropdown.js
@@ -1,9 +1,9 @@
 import Dropdown from "react-dropdown";
 import React from "react";
 import styled from 'styled-components';
-import {LANGS, THEMES} from "../renderers/Code";
-import {Labelled} from "../decorators/Labelled";
-import {Border, DropShadow, InputLike, Rounded} from "../Common/mixins";
+import { LANGS, THEMES } from "../renderers/Code";
+import { Labelled } from "../decorators/Labelled";
+import { Border, DropShadow, InputLike, Rounded } from "../Common/mixins";
 
 const StyledDropdown = styled(Dropdown)`
   ${Border}
